requirements.txt
fastapi>=0.110
uvicorn>=0.29
httpx>=0.27
openai>=1.14
pydantic>=2.6
python-dotenv>=1.0
